Addressed logic error in multi-master replication
Logic error resulted in an infinite loop. The ReplicationJudge was allowing the "LastTransactionReplicated" node to be replicated, resulting in an infinite replication loop. The solution was to add a check for the "LastTransactionReplicated" node label, and skip if found. 

Also addressed an issue preventing the replication procedure from working, by injecting the GraphDatabaseService into the procedure when it is called.